refactor: update compontent files to use kebab case
