+++ packages/outstatic/src/client/pages/media-library/media-library.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { AdminLayout } from '@/components'
+import { AdminLayout } from '@/components/admin-layout'
 import { API_MEDIA_PATH } from '@/utils/constants'
 import { useOutstatic } from '@/utils/hooks/useOutstatic'
 import { useGetMediaFiles } from '@/utils/hooks/useGetMediaFiles'
@@ -8,7 +8,7 @@ import { useState, useMemo } from 'react'
 import { toast } from 'sonner'
 import useSubmitMedia from '@/utils/hooks/useSubmitMedia'
 import { FileType } from '@/types'
-import { DeleteMediaButton } from '@/components/DeleteMediaButton'
+import { DeleteMediaButton } from '@/components/delete-media-button'
 import { MediaLibraryHeader } from '@/components/ui/outstatic/media-library-header'
 import { FileQuestion, ImageOff } from 'lucide-react'
 import { SpinnerIcon } from '@/components/ui/outstatic/spinner-icon'
