+++ packages/outstatic/src/components/delete-media-button.tsx
