--- packages/outstatic/src/components/index.ts
@@ -1,15 +1,0 @@
-import AdminHeader from './AdminHeader'
-import AdminLayout from './AdminLayout'
-import DocumentSettings from './DocumentSettings'
-import DocumentTitleInput from './DocumentTitleInput'
-import DocumentsTable from './DocumentsTable'
-import { Sidebar } from './Sidebar'
-
-export {
-  AdminHeader,
-  AdminLayout,
-  DocumentSettings,
-  DocumentTitleInput,
-  DocumentsTable,
-  Sidebar
-}
