+++ packages/outstatic/src/components/tests/delete-media-button.test.tsx
@@ -1,4 +1,4 @@
-import { DeleteMediaButton } from './index'
+import { DeleteMediaButton } from '@/components/delete-media-button'
 import { TestWrapper } from '@/utils/TestWrapper'
 import { fireEvent, render, screen, waitFor } from '@testing-library/react'
